Rename: `Complex` layout page -> `Default`
