--- src/modules/pages/components/RouteLayouts/Complex/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Complex';
