+++ src/modules/pages/components/RouteLayouts/Default/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Default';
